+++ portal-web/test/functional/com/liferay/portalweb/portal/selenium/SeleniumTestPlan.java
@@ -18,7 +18,7 @@ import com.liferay.portalweb.portal.BaseTestSuite;
 import com.liferay.portalweb.portal.selenium.assertions.AssertionsTestPlan;
 import com.liferay.portalweb.portal.selenium.browsercommands.BrowserCommandsTestPlan;
 import com.liferay.portalweb.portal.selenium.clicking.ClickingTestPlan;
-import com.liferay.portalweb.portal.selenium.select.SelectTestPlan;
+import com.liferay.portalweb.portal.selenium.list.ListTestPlan;
 import com.liferay.portalweb.portal.selenium.selection.SelectionTestPlan;
 import com.liferay.portalweb.portal.selenium.typing.TypingTestPlan;
 import com.liferay.portalweb.portal.selenium.waitfor.WaitForTestPlan;
@@ -37,7 +37,7 @@ public class SeleniumTestPlan extends BaseTestSuite {
 		testSuite.addTest(AssertionsTestPlan.suite());
 		testSuite.addTest(BrowserCommandsTestPlan.suite());
 		testSuite.addTest(ClickingTestPlan.suite());
-		testSuite.addTest(SelectTestPlan.suite());
+		testSuite.addTest(ListTestPlan.suite());
 		testSuite.addTest(SelectionTestPlan.suite());
 		testSuite.addTest(TypingTestPlan.suite());
 		testSuite.addTest(WaitForTestPlan.suite());
