+++ portal-web/test/functional/com/liferay/portalweb/portal/selenium/list/addselection/AddSelection2Test.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.selenium.select.addselection;
+package com.liferay.portalweb.portal.selenium.list.addselection;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,8 +20,8 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class AddSelection1Test extends BaseTestCase {
-	public void testAddSelection1() throws Exception {
+public class AddSelection2Test extends BaseTestCase {
+	public void testAddSelection2() throws Exception {
 		selenium.selectWindow("null");
 		selenium.selectFrame("relative=top");
 		selenium.open("/web/guest/home/");
@@ -51,8 +51,6 @@ public class AddSelection1Test extends BaseTestCase {
 			RuntimeVariables.replace("Select More Than One..."));
 		selenium.waitForVisible("//select[@id='_86_currentClassNameIds']");
 		selenium.addSelection("//select[@id='_86_currentClassNameIds']",
-			RuntimeVariables.replace("Blogs Entry"));
-		assertEquals("Blogs Entry",
-			selenium.getSelectedLabel("//select[@id='_86_currentClassNameIds']"));
+			RuntimeVariables.replace("Catherine"));
 	}
 }
