+++ portal-web/test/functional/com/liferay/portalweb/portal/selenium/list/addselection/AddSelection3Test.html
