+++ portal-web/test/functional/com/liferay/portalweb/portal/selenium/list/select/SelectTests.java
@@ -12,12 +12,9 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.selenium.select.addselection;
+package com.liferay.portalweb.portal.selenium.list.select;
 
-import com.liferay.portalweb.asset.assetpublisher.portlet.addportletap.AddPageAPTest;
-import com.liferay.portalweb.asset.assetpublisher.portlet.addportletap.AddPortletAPTest;
 import com.liferay.portalweb.portal.BaseTestSuite;
-import com.liferay.portalweb.portal.util.TearDownPageTest;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
@@ -25,15 +22,12 @@ import junit.framework.TestSuite;
 /**
  * @author Brian Wing Shun Chan
  */
-public class AddSelectionTests extends BaseTestSuite {
+public class SelectTests extends BaseTestSuite {
 	public static Test suite() {
 		TestSuite testSuite = new TestSuite();
-		testSuite.addTestSuite(AddPageAPTest.class);
-		testSuite.addTestSuite(AddPortletAPTest.class);
-		testSuite.addTestSuite(AddSelection1Test.class);
-		testSuite.addTestSuite(AddSelection2Test.class);
-		testSuite.addTestSuite(AddSelection3Test.class);
-		testSuite.addTestSuite(TearDownPageTest.class);
+		testSuite.addTestSuite(Select1Test.class);
+		testSuite.addTestSuite(Select2Test.class);
+		testSuite.addTestSuite(Select3Test.class);
 
 		return testSuite;
 	}
