+++ portal-web/test/functional/com/liferay/portalweb/portal/selenium/list/selectandwait/SelectAndWait1Test.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Select1Test</title>
+<title>SelectAndWait1Test</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">Select1Test</td></tr>
+<tr><td rowspan="1" colspan="3">SelectAndWait1Test</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
@@ -77,6 +77,11 @@
 	<td>selectAndWait</td>
 	<td>//select[@id='_128_add-permissions']</td>
 	<td>Sites</td>
+</tr>
+<tr>
+	<td>assertText</td>
+	<td>//form[@id='_128_fm']/h3</td>
+	<td>Page</td>
 </tr>
 </tbody>
 </table>
