+++ portal-web/test/functional/com/liferay/portalweb/portal/selenium/list/selectandwait/SelectAndWait1Test.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.selenium.select.selectandwait;
+package com.liferay.portalweb.portal.selenium.list.selectandwait;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,8 +20,8 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class SelectAndWait3Test extends BaseTestCase {
-	public void testSelectAndWait3() throws Exception {
+public class SelectAndWait1Test extends BaseTestCase {
+	public void testSelectAndWait1() throws Exception {
 		selenium.selectWindow("null");
 		selenium.selectFrame("relative=top");
 		selenium.open("/web/guest/home/");
@@ -51,8 +51,10 @@ public class SelectAndWait3Test extends BaseTestCase {
 		selenium.clickAt("link=Define Permissions",
 			RuntimeVariables.replace("Define Permissions"));
 		selenium.waitForPageToLoad("30000");
-		selenium.select("//select[@id='Catherine']",
+		selenium.select("//select[@id='_128_add-permissions']",
 			RuntimeVariables.replace("Sites"));
 		selenium.waitForPageToLoad("30000");
+		assertEquals(RuntimeVariables.replace("Page"),
+			selenium.getText("//form[@id='_128_fm']/h3"));
 	}
 }
