+++ portal-web/test/functional/com/liferay/portalweb/portal/selenium/list/selectandwait/SelectAndWait2Test.html
